feat: setup headless ui tabs
