+++ apps/web/src/components/InboxTabs/Seen.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const Seen = () => {
+  return (
+    <div>
+      <h3>SEEN</h3>
+    </div>
+  );
+};
