+++ apps/web/src/components/InboxTabs/Sent.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const Sent = () => {
+  return (
+    <div>
+      <h3>Sent</h3>
+    </div>
+  );
+};
