+++ apps/web/tailwind.config.js
@@ -21,7 +21,6 @@ module.exports = {
           200: '#1C1D21',
           300: '#111111',
           400: '#8E8E93',
-          border: '#505059',
         },
         dcblue: '#5865f2',
       },
